fit domain
